fix: 코드 리뷰 적용
