+++ package.json
@@ -35,7 +35,6 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
-    "embla-carousel-react": "^8.3.1",
     "fast-xml-parser": "^4.5.0",
     "lucide-react": "^0.453.0",
     "next": "14.2.15",
