+++ src/api/userApi.ts
@@ -1,4 +1,6 @@
+import { HopsitalItem } from "@/types/hospital";
 import { SupabaseDatabase } from "@/types/supabaseDataType";
+import browserClient from "@/utils/supabase/client";
 
 // supabase에서 로그인 정보 가져오기
 export const getUser = async (supabaseClient: SupabaseDatabase) => {
@@ -31,11 +33,14 @@ export const getChildren = async (supabaseClient: SupabaseDatabase, userId?: str
   }
 };
 
-
 // like table에서 사용자가 좋아요한 병원 목록 가져오기
 export const getUserLike = async (supabaseClient: SupabaseDatabase, userId: string) => {
   if (userId) {
-    const { data, error } = await supabaseClient.from("like").select().eq("user_id", userId);
+    const { data, error } = await supabaseClient
+      .from("like")
+      .select()
+      .eq("user_id", userId)
+      .order("created_at", { ascending: false }); // 최근 등록 순
 
     if (error) {
       console.error(error.message);
@@ -45,5 +50,29 @@ export const getUserLike = async (supabaseClient: SupabaseDatabase, userId: stri
     return data;
   } else {
     return [];
+  }
+};
+
+export const addLike = async (hospitalInfo: HopsitalItem) => {
+  const {
+    orgnm,
+    orgcd,
+    orgAddr,
+    orgTlno,
+    expnYmd,
+    vcnList: { vcnInfo: tmpInfo }
+  } = hospitalInfo;
+  const vcnInfo = Array.isArray(tmpInfo) ? tmpInfo : [tmpInfo];
+  const hospitalData = { orgnm, orgcd, orgAddr, orgTlno, expnYmd, vcnInfo: JSON.stringify(vcnInfo) };
+  const { error } = await browserClient.from("like").insert(hospitalData);
+
+  if (error) throw Error(error.message);
+};
+
+export const cancelLike = async (id?: string) => {
+  if (id) {
+    const { error } = await browserClient.from("like").delete().eq("id", id);
+
+    if (error) throw Error(error.message);
   }
 };
